requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
Pillow>=9.0.0
matplotlib>=3.7.0
rasterio>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
folium>=0.14.0
streamlit-folium>=0.15.0
geopandas>=0.14.0
scikit-learn>=1.3.0
